+++ src/app/store/reducers/pagination-reducer/pagination-reducer-clear-pagination-type.ts
@@ -1,0 +1,9 @@
+import { PaginationState } from '../../types/pagination.types';
+export function paginationClearType(state: PaginationState, action) {
+  if (state[action.entityKey]) {
+    const clearState = { ...state };
+    clearState[action.entityKey] = {};
+    return clearState;
+  }
+  return state;
+}
