+++ src/app/store/reducers/pagination-reducer/pagination.reducer.ts
@@ -1,41 +1,28 @@
-import { resultPerPageParam, resultPerPageParamDefault } from './pagination-reducer.types';
-import {
-  getActionKey,
-  getActionType,
-  getPaginationKey,
-  getUniqueQParams,
-  removeEmptyParams,
-} from './pagination-reducer.helper';
-import { PaginatedAction, PaginationEntityState, PaginationParam, QParam } from '../../types/pagination.types';
 import { error } from 'util';
-import { Action, Store } from '@ngrx/store';
-import { denormalize, Schema } from 'normalizr';
 
-import { ApiActionTypes } from '../../actions/request.actions';
 import {
   ADD_PARAMS,
-  AddParams,
   CLEAR_PAGES,
   CLEAR_PAGINATION_OF_TYPE,
   REMOVE_PARAMS,
-  RemoveParams,
   SET_PAGE,
   SET_PARAMS,
-  SetPage,
-  SetParams,
 } from '../../actions/pagination.actions';
-import { AppState } from '../../app-state';
+import { ApiActionTypes } from '../../actions/request.actions';
 import { mergeState } from '../../helpers/reducer.helper';
-import { Observable } from 'rxjs/Observable';
-import { selectPaginationState } from '../../selectors/pagination.selectors';
 import { defaultCfEntitiesState } from '../../types/entity.types';
+import { PaginationEntityState, PaginationState } from '../../types/pagination.types';
+import { paginationAddParams } from './pagination-reducer-add-params';
+import { paginationClearPages } from './pagination-reducer-clear-pages';
+import { paginationClearType } from './pagination-reducer-clear-pagination-type';
+import { paginationRemoveParams } from './pagination-reducer-remove-params';
+import { paginationSetPage } from './pagination-reducer-set-page';
+import { paginationSetParams } from './pagination-reducer-set-params';
+import { paginationStart } from './pagination-reducer-start';
 import { paginationSuccess } from './pagination-reducer-success';
 import { paginationFailure } from './pagination-reducer.failure';
-import { paginationStart } from './pagination-reducer-start';
-import { paginationSetParams } from './pagination-reducer-set-params';
-import { paginationSetPage } from './pagination-reducer-set-page';
-import { paginationAddParams } from './pagination-reducer-add-params';
-import { paginationRemoveParams } from './pagination-reducer-remove-params';
+import { getActionKey, getActionType, getPaginationKey } from './pagination-reducer.helper';
+import { resultPerPageParam, resultPerPageParamDefault } from './pagination-reducer.types';
 
 const defaultPaginationEntityState = {
   fetching: false,
@@ -76,8 +63,6 @@ const getPaginationUpdater = function (types: [string, string, string]) {
   };
 };
 
-
-
 export function createPaginationReducer(types: [string, string, string]) {
   const updatePagination = getPaginationUpdater(types);
   const [requestType, successType, failureType] = types;
@@ -88,51 +73,32 @@ export function createPaginationReducer(types: [string, string, string]) {
     }
 
     if (action.type === CLEAR_PAGES) {
-      if (state[action.entityKey] && state[action.entityKey][action.paginationKey]) {
-        const newState = { ...state };
-        const entityState = {
-          ...newState[action.entityKey],
-          [action.paginationKey]: {
-            ...newState[action.entityKey][action.paginationKey],
-            ids: {},
-            fetching: false,
-            pageCount: 0,
-            currentPage: 1,
-            totalResults: 0,
-            error: false,
-            message: ''
-          }
-        };
-        return {
-          ...newState,
-          [action.entityKey]: entityState
-        };
-      }
+      return paginationClearPages(state, action);
     }
 
     if (action.type === CLEAR_PAGINATION_OF_TYPE) {
-      if (state[action.entityKey]) {
-        const clearState = { ...state };
-        clearState[action.entityKey] = {};
-        return clearState;
-      }
-      return state;
+      return paginationClearType(state, action);
     }
 
-    const actionType = getActionType(action);
-    const key = getActionKey(action);
-    const paginationKey = getPaginationKey(action);
-    if (actionType && key && paginationKey) {
-      const newState = { ...state };
-      const updatedPaginationState = updatePagination(newState[key][paginationKey], action, actionType);
-      newState[key] = mergeState(newState[key], {
-        [paginationKey]: updatedPaginationState
-      });
-      return newState;
-    } else {
-      return state;
-    }
+    return enterPaginationReducer(state, action, updatePagination);
   };
+}
+
+function enterPaginationReducer(state: PaginationState, action, updatePagination) {
+  const actionType = getActionType(action);
+  const key = getActionKey(action);
+  const paginationKey = getPaginationKey(action);
+
+  if (actionType && key && paginationKey) {
+    const newState = { ...state };
+    const updatedPaginationState = updatePagination(newState[key][paginationKey], action, actionType);
+    newState[key] = mergeState(newState[key], {
+      [paginationKey]: updatedPaginationState
+    });
+    return newState;
+  } else {
+    return state;
+  }
 }
